# Avoid calling addNulls in exprEval when there is no nulls to add.