Summary:
before:
```
============================================================================
[...]lox/benchmarks/DenseProcBenchmark.cpp     relative  time/iter   iters/s
============================================================================
koskiExpandedOnVeloxIfFloor                               283.95ns     3.52M
koskiExpandedOnVeloxOneHot                                219.04ns     4.57M
koskiExpandedOnVeloxOneHot_OneHotHitting                  216.28ns     4.62M
```

after:
```
============================================================================
[...]lox/benchmarks/DenseProcBenchmark.cpp     relative  time/iter   iters/s
============================================================================
koskiExpandedOnVeloxIfFloor                               266.17ns     3.76M
koskiExpandedOnVeloxOneHot                                190.80ns     5.24M
koskiExpandedOnVeloxOneHot_OneHotHitting                  192.89ns     5.18M
```

Differential Revision: D41660725